Share the parquet cache staleness rule between the estimators and the packets
The packets reader held a second copy of the rule that estimators.py had
already factored out as rankbatch_parquet_staleness: the same call to
read_parquet_cache_metadata, the same one-directional comparison for an
incomplete batch, the same ValueError fallback, and the same message.

Thus a change to the tolerance, a new metadata key, or a correction to the
incomplete-batch rule reached one reader alone. A user who restarts a run and
rewrites some packets00_*.out files then got stale packet data from the cache
while the estimators of the same run were correctly rebuilt.

The rule now lives beside read_parquet_cache_metadata in misc/fileio.py and
takes the cache format version as an argument, because the two caches version
their formats on their own.
