+++ artistools/estimators/test_estimators.py
@@ -786,14 +786,17 @@ def test_a_current_parquet_cache_starts_no_progress_bar(tmp_path: Path) -> None:
     from artistools.estimators.estimators import CACHEVERSION
     from artistools.estimators.estimators import get_rankbatch_parquetpath
     from artistools.estimators.estimators import rankbatch_parquet_is_current
-    from artistools.estimators.estimators import rankbatch_parquet_staleness
     from artistools.misc.fileio import MTIME_TOLERANCE_S
+    from artistools.misc.fileio import rankbatch_parquet_staleness
 
     parquetfilepath = get_rankbatch_parquetpath(tmp_path, [0, 1, 2], 0)
     assert parquetfilepath.name == "estimbatch00_0000_0002.out.parquet.tmp"
 
     # a cache that no run wrote yet needs the conversion
-    assert rankbatch_parquet_staleness(parquetfilepath, None, textsource_complete=False) == "the file does not exist"
+    assert (
+        rankbatch_parquet_staleness(parquetfilepath, CACHEVERSION, None, textsource_complete=False)
+        == "the file does not exist"
+    )
 
     mtime = 1000.0
     at.write_parquet_atomic(
@@ -2081,9 +2084,10 @@ def test_an_archived_run_keeps_a_cache_of_an_old_version(tmp_path: Path) -> None
     No conversion can replace such a cache. A rejection made get_runfolder_timesteps() find no
     timesteps, thus get_runfolders() dropped the folder and the reader saw a run that holds no data.
     """
+    from artistools.estimators.estimators import CACHEVERSION
     from artistools.estimators.estimators import rankbatch_cache_cannot_be_rebuilt
     from artistools.estimators.estimators import rankbatch_parquet_is_current
-    from artistools.estimators.estimators import rankbatch_parquet_staleness
+    from artistools.misc.fileio import rankbatch_parquet_staleness
 
     oldversion = tmp_path / "estimbatch00_0000_0002.out.parquet.tmp"
     at.write_parquet_atomic(
@@ -2091,7 +2095,9 @@ def test_an_archived_run_keeps_a_cache_of_an_old_version(tmp_path: Path) -> None
     )
 
     # the reason still names the fault, so that the reader can give a warning
-    assert "cache format version" in str(rankbatch_parquet_staleness(oldversion, None, textsource_complete=False))
+    assert "cache format version" in str(
+        rankbatch_parquet_staleness(oldversion, CACHEVERSION, None, textsource_complete=False)
+    )
     # but the cache stays readable, because no conversion can replace it
     assert rankbatch_cache_cannot_be_rebuilt(oldversion, textsource_complete=False)
     assert rankbatch_parquet_is_current(oldversion, None, textsource_complete=False)
@@ -2171,10 +2177,13 @@ def test_a_batch_that_keeps_only_rank_zero_keeps_its_cache(tmp_path: Path) -> No
     # a rewrite of the file that remains still proves that the cache is stale
     os.utime(tmp_path / "estimators_0000.out", (1400.0, 1400.0))
     mtime_rewritten, complete_rewritten = get_batch_textsource_state(get_textsource_mtimes(tmp_path), 0, 2)
-    from artistools.estimators.estimators import rankbatch_parquet_staleness
+    from artistools.estimators.estimators import CACHEVERSION
+    from artistools.misc.fileio import rankbatch_parquet_staleness
 
     assert "after the cache stamp" in str(
-        rankbatch_parquet_staleness(parquetfilepath, mtime_rewritten, textsource_complete=complete_rewritten)
+        rankbatch_parquet_staleness(
+            parquetfilepath, CACHEVERSION, mtime_rewritten, textsource_complete=complete_rewritten
+        )
     )
 
 
@@ -2263,8 +2272,9 @@ def test_a_partial_batch_keeps_a_cache_of_an_old_version(tmp_path: Path) -> None
     read the one text file that remains, thus it lost the timesteps that only the cache holds and it
     named a set that the scan cannot deliver.
     """
+    from artistools.estimators.estimators import CACHEVERSION
     from artistools.estimators.estimators import rankbatch_parquet_is_current
-    from artistools.estimators.estimators import rankbatch_parquet_staleness
+    from artistools.misc.fileio import rankbatch_parquet_staleness
     from artistools.misc.modelinfo import get_runfolder_timesteps
 
     runfolder = tmp_path / "job1.slurm"
@@ -2283,7 +2293,7 @@ def test_a_partial_batch_keeps_a_cache_of_an_old_version(tmp_path: Path) -> None
 
     # the reason still names the fault, so that the reader can give a warning
     assert "cache format version" in str(
-        rankbatch_parquet_staleness(parquetfilepath, 1000.0, textsource_complete=False)
+        rankbatch_parquet_staleness(parquetfilepath, CACHEVERSION, 1000.0, textsource_complete=False)
     )
     # but the cache answers, because the scan of the run keeps it for the same reason
     assert rankbatch_parquet_is_current(parquetfilepath, 1000.0, textsource_complete=False)
