+++ artistools/packets/packets.py
@@ -18,11 +18,9 @@ from artistools.constants import C_cm_per_s as CLIGHT
 from artistools.constants import day_to_s
 from artistools.constants import km_to_cm
 from artistools.misc import print_warning
-from artistools.misc import read_parquet_cache_metadata
 from artistools.misc.fileio import COMPRESSED_EXTENSIONS
-from artistools.misc.fileio import format_mtime
-from artistools.misc.fileio import MTIME_TOLERANCE_S
 from artistools.misc.fileio import parquet_is_readable
+from artistools.misc.fileio import rankbatch_parquet_staleness
 
 type_ids = {"TYPE_GAMMA": 10, "TYPE_RPKT": 11, "TYPE_NTLEPTON": 20, "TYPE_ESCAPE": 32}
 
@@ -415,27 +413,13 @@ def get_packets_rankbatch_parquetfile(
         # modification times give no full comparison, but the cache format version still applies
         textsource_mtime = max(textsource_mtimes) if allranksfound else None
 
-        # an archived run of packets costs hours to convert again, thus a cache from before the stamps
-        # stays in use. See the accept_unstamped argument of read_parquet_cache_metadata
-        pqmetadata, stalereason = read_parquet_cache_metadata(
-            parquetfilepath, CACHEVERSION, textsource_mtime, accept_unstamped=True
+        # one rule decides the freshness of every batch cache of this repository
+        stalereason = rankbatch_parquet_staleness(
+            parquetfilepath,
+            CACHEVERSION,
+            max(textsource_mtimes) if textsource_mtimes else None,
+            textsource_complete=allranksfound,
         )
-
-        if stalereason is None and not allranksfound and textsource_mtimes:
-            # a text file that is newer than the stamp proves that a restart rewrote the source after
-            # the cache. An absent text file gives no proof, thus only this one test applies
-            stampedmtime = (pqmetadata or {}).get("textsource_mtime")
-            try:
-                # a stamp that no writer of this repository can produce dates no text file
-                stampedvalue = float(stampedmtime) if stampedmtime is not None else None
-            except ValueError:
-                stampedvalue = None
-
-            if stampedvalue is not None and max(textsource_mtimes) > stampedvalue + MTIME_TOLERANCE_S:
-                stalereason = (
-                    f"a text file of the batch changed at {format_mtime(max(textsource_mtimes))},"
-                    f" after the cache stamp of {format_mtime(stampedmtime)}"
-                )
 
         if stalereason is None:
             conversion_needed = False
@@ -452,7 +436,7 @@ def get_packets_rankbatch_parquetfile(
                 f" ranks {batch_mpiranks[0]} to {batch_mpiranks[-1]}, because {stalereason}."
                 " File will be regenerated..."
             )
-        elif pqmetadata is not None or parquet_is_readable(parquetfilepath):
+        elif parquet_is_readable(parquetfilepath):
             # the text files are incomplete, thus no conversion can replace this cache. The data can
             # be older than the text files that remain, thus the user needs a warning
             conversion_needed = False
